refs #78677: fix linting errors
