+++ apps/src/components/map-layers/map-events.ts
@@ -23,7 +23,7 @@ export default function MapEvents({
 				onUnmount();
 			}
 		};
-	}, [map, onMapReady]);
+	}, [map, onMapReady, onUnmount]);
 
 	return null;
 }
