+++ apps/src/lib/format.ts
@@ -1,3 +1,4 @@
+import { cva } from 'class-variance-authority';
 import {
 	WMSLegendData,
 	TransformedLegendEntry,
@@ -92,13 +93,13 @@ export const normalizeRadioCardProps = async <T extends 'taxonomy' | 'post'>(
 
 			const taxonomies = terms.reduce(
 				// TODO: replace with correctly typed params
-				(acc: Record<string, any[]>, term: { taxonomy: string }) => {
+				(acc: Record<string, unknown[]>, term: { taxonomy: string }) => {
 					if (term.taxonomy) {
 						acc[term.taxonomy] = postItem[term.taxonomy] || [];
 					}
 					return acc;
 				},
-				{} as Record<string, any[]>
+				{} as Record<string, unknown[]>
 			);
 
 			return {
@@ -118,3 +119,33 @@ export const normalizeRadioCardProps = async <T extends 'taxonomy' | 'post'>(
 
 	return normalizedData as T extends 'taxonomy' ? TaxonomyData[] : PostData[];
 };
+
+export const buttonVariants = cva(
+	'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0',
+	{
+		variants: {
+			variant: {
+				default:
+					'bg-primary text-primary-foreground hover:bg-primary/90',
+				destructive:
+					'bg-destructive text-destructive-foreground hover:bg-destructive/90',
+				outline:
+					'border border-input bg-background hover:bg-accent hover:text-accent-foreground',
+				secondary:
+					'bg-secondary text-secondary-foreground hover:bg-secondary/80',
+				ghost: 'hover:bg-accent hover:text-accent-foreground',
+				link: 'text-primary underline-offset-4 hover:underline',
+			},
+			size: {
+				default: 'h-10 px-4 py-2',
+				sm: 'h-9 rounded-md px-3',
+				lg: 'h-11 rounded-md px-8',
+				icon: 'h-10 w-10',
+			},
+		},
+		defaultVariants: {
+			variant: 'default',
+			size: 'default',
+		},
+	}
+);
